# [GEOT-6900] Shapefile quadtree build performance improvements