… modules/plugin/shapefile/src/test/java/org/geotools/data/shapefile/ShapeFileIndexerStressTest.java