@groldan 2021-06-02
Amended to fix merge conflicts and respect the new PMD rules